Modify scombz.js
