+++ src/js/scombz/improveExamView.js
@@ -1,0 +1,51 @@
+export default () => {
+  // テスト改善
+  if (
+    location.href.indexOf('examination') !== -1 &&
+    document.body.clientWidth > 480
+  ) {
+    if (s2b($settings_exam)) {
+      console.log('テスト改善を実行します')
+      const $exa_contsize = document.getElementById('pageContents')
+      const $exa_examImgList = document.querySelectorAll('.downloadFile')
+      const $exa_cheadList = document.querySelectorAll('.contents-header')
+      const $exa_img = document.querySelector('.exam-question-img')
+      const $exa_footer = document.getElementById('page_foot')
+      const $exa_timer = document.getElementById('examTimer')
+      if ($exa_footer) {
+        $exa_footer.style.visibility = 'hidden'
+      }
+
+      for (const $exa_chead of $exa_cheadList) {
+        $exa_chead.style.width = '8%'
+        $exa_chead.style.background = '#f6f6ff'
+      }
+      if ($exa_img) {
+        for (const $exa_examImg of $exa_examImgList) {
+          $exa_examImg.style.maxHeight = '100vh'
+          $exa_examImg.style.boxShadow = '0 0 1px #000000 '
+        }
+        $exa_examImgList[0].style.maxHeight = '95vh'
+        $exa_examImgList[0].style.maxWidth = '50vw'
+        $exa_examImgList[0].style.position = 'fixed'
+        $exa_examImgList[0].style.right = '1px'
+        $exa_examImgList[0].style.top = '5vh'
+        $exa_examImgList[0].style.boxShadow = '0 0 1px #000000 '
+        if ($exa_contsize) {
+          $exa_contsize.style.width =
+            document.body.clientWidth -
+            $exa_examImgList[0].clientWidth -
+            3 +
+            'px'
+        }
+        if ($exa_timer) {
+          $exa_timer.style.width =
+            document.body.clientWidth - $exa_examImgList[0].clientWidth + 'px'
+        }
+      }
+      console.log('テスト改善の実行が完了しました')
+    }
+  } else {
+    // Nothing
+  }
+}
